Added OptionPanel3, an update version of OptionPanel2.
This OptionPanel3 is still HALF-IMPLEMENTED, however, it works in a
different way from that of OptionPanel2. OptionPanel3 generally works
with bindings and converters, in other words, it got rid of updating
option sources via "*Changed"events of the UI elements. In addition, a
public method UpdateOptionsUI is implemented for updating binding target
values after initialization, so CALL IT IN MAINWINDOW or something after
parsing and loading the skins.